requirements.txt
python-dotenv==1.0.1
pyTelegramBotAPI==4.25.0
requests==2.32.3